pyproject.toml
[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 110
fix = true
select = ["E", "F", "W", "C", "D", "I", "N", "Q"]
ignore = ["D203", "D213", "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"

[tool.ruff.isort]
combine-as-imports = true

[tool.ruff.per-file-ignores]
"*/tests/*.py" = ["D", "E501"]
"*/migrations/*.py" = ["ALL"]
[project]
name = "nfac-hws"
version = "0.1.0"
description = ""
authors = [
    {name = "Niyaz",email = "Niyaz@mail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0"
alembic = "^1.13"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
pydantic = "^2.0"
python-jose = "^3.3"
passlib = {extras = ["bcrypt"], version = "^1.7"}
loguru = "^0.7.3"
